Custom Hooks Implementation
PnPJs Team was looking for a new version of the react-pnpjs-sample when using custom hooks.
